+++ INB348/GroupsRequestListTableViewController.m
@@ -24,7 +24,7 @@
         [query findObjectsInBackgroundWithBlock:^(NSArray *objects, NSError *error) {
             if (!error) {
                 // The find succeeded.
-                NSLog(@"Successfully retrieved %d UserGroups.", objects.count);
+                NSLog(@"Successfully retrieved %lu UserGroups.", (unsigned long)objects.count);
                 
                 // Do something with the found objects
                 self.requestedGroupUsers = [objects mutableCopy];
